+++ docs/conf.py
@@ -145,6 +145,8 @@ nitpick_ignore = [
     (_PY_CLASS, 'dmr.decorators._ReturnT'),
     (_PY_CLASS, 'dmr.decorators._ViewT'),
     (_PY_CLASS, 'dmr.decorators._TypeT'),
+    (_PY_CLASS, 'dmr.internal.routing.URLExternal'),
+    (_PY_CLASS, '_URLExternal'),
     (_PY_CLASS, 'dmr.internal.negotiation.ConditionalType'),
     (_PY_CLASS, 'dmr.security.jwt.views._ObtainTokensT'),
     (_PY_CLASS, 'dmr.security.jwt.views._RefreshTokensT'),